requirements.txt
streamlit
selenium
pandas
langchain
google-generativeai
openai

